run: Move in_color from run.py to helper.py.
This is the right place for `in_color` as it can be used to show
custom colored messages in terminal across zt.
